refactor: update folders, move and rename files
