+++ src/components/client/Components/ResponseSection/Statistic/index.tsx
